factor out types
